Made option optional work with multi hints
Option optional was designed in combination with type option so as
to take into account those props that may not always be set but
should be used when they are.

On the other hand, option ignore was designed to overlook certain
arguments in keyfunc altoghether, yielding to their always being
ignored. Now option optional can be used instead.